Update Query_under_5_group_by.sql
Make it so that WHERE uses float value instead of dollar, and also filters out categories that are not choc or toffee. Also make it so that headers show in query output, as well as changing the COUNT as sweet_count using AS. Now the query will only show the count of the chocolate and toffees that are under 5 dollars, which is what the task goal was.